e: systemd-wrapped e17 start

+++ xinitrc.e
@@ -2,22 +2,20 @@
 
 # Basic X setup
 xrdb ~/.Xresources
-export XDG_DATA_DIRS=$HOME/.xdg:/usr/share/enlightenment:/usr/share
-
-# DBus session bus
-[ -z "$DBUS_SESSION_BUS_ADDRESS" ] &&\
-	eval $(dbus-launch --sh-syntax --exit-with-session)
 
 # Agents
-pkill gpg-agent
-pkill ssh-agent
-truncate -s0 /dev/shm/zenv_agents 2>/dev/null
-chgrp wheel /dev/shm/zenv_agents 2>/dev/null
-chmod 660 /dev/shm/zenv_agents 2>/dev/null
-gpg-agent >>/dev/shm/zenv_agents --daemon --pinentry-program $(which pinget)
-ssh-agent | grep -v '^echo' >>/dev/shm/zenv_agents
+if ! pgrep gpg-agent || ! pgrep ssh-agent; then
+	pkill gpg-agent
+	pkill ssh-agent
+	truncate -s0 /dev/shm/zenv_agents 2>/dev/null
+	chgrp wheel /dev/shm/zenv_agents 2>/dev/null
+	chmod 660 /dev/shm/zenv_agents 2>/dev/null
+	gpg-agent >>/dev/shm/zenv_agents --daemon --pinentry-program $(which pinget)
+	ssh-agent | grep -v '^echo' >>/dev/shm/zenv_agents
+fi
 source /dev/shm/zenv_agents
 
-# e17/systemd session
-ck-launch-session enlightenment_start
-pkill -INT -U $UID -x systemd ||: # make sure it's dead
+# systemd/e17 session
+export XDG_DATA_DIRS="${HOME}/.xdg:/usr/share/enlightenment:/usr/share"
+export DBUS_SESSION_BUS_ADDRESS="unix:path=${XDG_RUNTIME_DIR}/session_bus_socket"
+ck-launch-session /usr/lib/systemd/systemd --user
